separate Ice Schema

+++ controllers/ticketController.js
@@ -1,10 +1,12 @@
+import mongoose from "mongoose";
 import Ticket from "../models/ticketModel.js";
 import Ice from "../models/iceModel.js";
 
 //  ! get all tickets
 export const getAllTickets = async (req, res) => {
   try {
-    const tickets = await Ticket.find();
+    const tickets = await Ticket.find().populate("ice")
+    // .populate("author");
     res.status(200).json(tickets);
   } catch (err) {
     res.status(404).json({ message: err.message });
@@ -25,7 +27,8 @@ export const getSubmittedTickets = async (req, res) => {
 export const getSingleTicket = async (req, res) => {
   const { id } = req.params;
   try {
-    const ticket = await Ticket.findById(id);
+    const ticket = await Ticket.findById(id).populate("ice")
+    // .populate("author");
     res.status(200).json(ticket);
   } catch (err) {
     res.status(404).json({ message: err.message });
@@ -34,8 +37,24 @@ export const getSingleTicket = async (req, res) => {
 
 // ! create new Ticket
 export const createTicket = async (req, res) => {
-  try {
-    const newTicket = await Ticket.create(req.body);
+ 
+  try { 
+    // * create new Ice
+    const newIce = await Ice.create({
+      impact: req.body.impact,
+      effort: req.body.effort,
+      confidence: req.body.confidence,
+    });
+
+    // * after ice save use its _id as ObjectId to ticket get Ice and Ticket Data nested
+    const newTicket = await Ticket.create({
+      initialtive: req.body.initialtive,
+      description: req.body.description,
+      ice: newIce._id,
+      isSubmitted: req.body.isSubmitted,
+      target: req.body.target,
+      ticket_id: req.body.ticket_id
+    });
 
     res.status(200).json(newTicket);
   } catch (err) {
@@ -60,18 +79,31 @@ export const createTicket = async (req, res) => {
 
 // ! update Ticket
 // ? need to be efficent (Use 2 find method)
-
 export const updateTicket = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).send(`No ticket with id: ${id}`);
   const ticket = await Ticket.findById(id);
-  
-  await Ticket.findByIdAndUpdate(id, req.body, { new: true });
-  res.json(req.body);
+  const updateTicket = {
+    initialtive: req.body.initialtive,
+    description: req.body.description,
+    ice: {
+      impact: req.body.impact,
+      confidence: req.body.confidence,
+      effort: req.body.effort,
+      _id: ticket.ice,
+    },
+    _id: id,
+  };
+  await Ticket.findByIdAndUpdate(id, updateTicket, { new: true });
+  res.json(updateTicket);
 };
 
 // ! delet Ticket
 export const deleteTicket = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).send(`No post with id: ${id}`);
   await Ticket.findByIdAndRemove(id);
   res.json({ message: "Ticket deleted successfully" });
 };
